move fetch_backend
