+++ frontend/app/novels/novels-list/upload-backup.tsx
@@ -16,7 +16,7 @@ import { ChangeEvent, FormEvent, useState } from "react"
 import { cn } from "@/lib/utils"
 import { buttonVariants } from "@/components/ui/button"
 
-import { fetch_backend } from "@/utils/fetch_backend"
+import { fetch_backend } from "@/lib/fetch_backend"
 
 export function UploadBackupDialog() {
   const [file, setFile] = useState<File | null>(null);
